+++ tests/fixtures/fail/shorthand-tag.php
@@ -1,0 +1,6 @@
+<div
+	class="example <?= esc_attr( $class ); ?>"
+	data-action="<?= esc_attr( $action ); ?>"
+>
+	<?= esc_html( $content ); ?>
+</div>